PCB edge integration, supported at mounting holes
PCB edge is parsed from Edge Cuts layer. Convex
hull determines the border of a support structure.
This has to be inset to support the board. And
offset to create a slot for the board to sit.
Offset further yet, and extruded out to form the
lip. Essentially the board sits in a slot.

Mounting holes are taken, and only plastic near that
is retained, intersecting with the previous structure.