Connect delaunay mesh to the PCB edge.
Corners of bounding box of the PCB are now included in
delaunay triangulation. Thus, they can't form a separate
mesh - and will be connected to the PCB edge irrespective
of PCB design.

This also opens up the chance of the delaunay triangles
lying outside the PCB edge. If the PCB edge is say a circle,
this can happen.  Thus, the delaunay triangles must be
intersected with the "external volume" of the PCB - that
is the volume enclosed in the external perimeter.