Robust handling of 3d models in footprints
Each 3d model in every footprint (TH) is now processed properly.
It is transformed in 3D space according to its properties. Then
checked whether it intersects the board. If not, then the model
is ignored.  This is important for compound footprints like the
Pico or the mini 360. For visualization purposes, entire models
are included in the board file.  However, only the berg headers
are soldered on the board.  Rest must not be considered.

Shells can also protrude into each other, given enough thickness,
or depending on the placement.  The volume of the components
("pads") must not have any walls coming there - so all the pads
are subtracted again at the end. This difference() operation
cuts into the PCB edge as well, if that is the right thing to do.