Ensure 3D models from /usr/share/kicad are found
KICAD[6-8]_3DMODEL_DIR are defined.  I have KICAD8. I see many
models use the old version identifiers. There might be good
reasons for that.  In any case, we look at the environment,
and define stuff that is undefined. The user thus always has
a choice if they way to point to somewhere else.