Experiment: "hugging" shells, based on z binning.
Existing method of single shells allows components to
wiggle in ways that make it difficult to insert a PCB.
e.g. single row berg headers can wiggle on the long
axis, as there is space available.  Also, placing
them precisely vertical isn't easy, and due to the wiggle
room, they don't stay that way. This makes it hard
to insert the PCB in the first time, when the components
are put into their holding shells in the entire jig.

Here, we work on the idea we need to create pockets for
every "level", and with that we can ensure precise
placement and almost eliminate wiggling

This has worked very well. I have tested a 1x5 and
and a 2x5 standard 2.54 mm berg header.

Also tested a single pin 2.54 mm berg header. Found
something interesting here. The area of the pin
is about 0.4096 mm2, making a 0.64 mm x 0.64 mm
hole. My printer couldn't reproduce this. So the
pin didn't go in. It make a tiny hole though! I didn't
pursue widening the hole as the jig shouldn't need
manual postprocessing.

I added offsets of 0.3, 0.5 and 0.7 mm. The 0.3 mm
offset model fit, and so did the rest. Some some
logic must be added in the tools.