Use SolidPython2 to generate openscad, mostly
SolidPython2 is a cool library to generate openscad files
pythonically. It doesn't support a native "module" yet. This
is included in solid2_module.py

The usage of SolidPython2 should be streamlined even more.
But this is a reasonable start. I haven't yet eliminated
scad code generation entirely, as there are still benefits
to print statements.  E.g. it would make little sense for
SolidPython to support "if" statements in scad. But we need
them to offer more flexibility of usage of generated files
in openscad.